Move test and funcs, return only role names
