+++ tool/tctl/common/accesslist/preset.go
@@ -25,6 +25,7 @@ import (
 	"github.com/gravitational/teleport/api/types/common"
 	"github.com/gravitational/teleport/lib/accesslists/preset"
 	"github.com/gravitational/teleport/lib/utils"
+	"github.com/gravitational/teleport/lib/utils/parse"
 )
 
 const (
@@ -38,6 +39,120 @@ const (
 )
 
 var awsIcAppLabel = types.Labels{types.OriginLabel: []string{common.OriginAWSIdentityCenter}}
+
+type applyAccessFlagsToRole func(allow *types.RoleConditions) error
+
+// applyStandardAccessFlagsToRole modifies the standard role's allow block.
+// Empty values clear the field, unset flags leave field alone.
+func (c *Command) applyStandardAccessFlagsToRole(allow *types.RoleConditions) error {
+	// Nodes
+	if c.nodeLabelsSet {
+		labels, err := parse.MultiValueLabelSelectorSpec(c.nodeLabels)
+		if err != nil {
+			return trace.Wrap(err, "--node-labels")
+		}
+		allow.NodeLabels = types.ToLabels(labels)
+	}
+	if c.loginsSet {
+		allow.Logins = utils.SplitIdentifiers(c.logins)
+	}
+
+	// Dbs
+	if c.dbLabelsSet {
+		labels, err := parse.MultiValueLabelSelectorSpec(c.dbLabels)
+		if err != nil {
+			return trace.Wrap(err, "--db-labels")
+		}
+		allow.DatabaseLabels = types.ToLabels(labels)
+	}
+	if c.dbUsersSet {
+		allow.DatabaseUsers = utils.SplitIdentifiers(c.dbUsers)
+	}
+	if c.dbNamesSet {
+		allow.DatabaseNames = utils.SplitIdentifiers(c.dbNames)
+	}
+
+	// Kubes
+	if c.kubeLabelsSet {
+		labels, err := parse.MultiValueLabelSelectorSpec(c.kubeLabels)
+		if err != nil {
+			return trace.Wrap(err, "--kubernetes-labels")
+		}
+		allow.KubernetesLabels = types.ToLabels(labels)
+	}
+	if c.kubeUsersSet {
+		allow.KubeUsers = utils.SplitIdentifiers(c.kubeUsers)
+	}
+	if c.kubeGroupsSet {
+		allow.KubeGroups = utils.SplitIdentifiers(c.kubeGroups)
+	}
+
+	// Apps
+	if c.appLabelsSet {
+		labels, err := parse.MultiValueLabelSelectorSpec(c.appLabels)
+		if err != nil {
+			return trace.Wrap(err, "--app-labels")
+		}
+		allow.AppLabels = types.ToLabels(labels)
+	}
+	if c.awsRoleARNsSet {
+		allow.AWSRoleARNs = utils.SplitIdentifiers(c.awsRoleARNs)
+	}
+	if c.azureIdentitiesSet {
+		allow.AzureIdentities = utils.SplitIdentifiers(c.azureIdentities)
+	}
+	if c.gcpServiceAccountsSet {
+		allow.GCPServiceAccounts = utils.SplitIdentifiers(c.gcpServiceAccounts)
+	}
+	if c.mcpToolsSet {
+		tools := utils.SplitIdentifiers(c.mcpTools)
+		if len(tools) == 0 {
+			allow.MCP = nil
+		} else {
+			allow.MCP = &types.MCPPermissions{Tools: tools}
+		}
+	}
+
+	// Windows
+	if c.windowsLabelsSet {
+		labels, err := parse.MultiValueLabelSelectorSpec(c.windowsLabels)
+		if err != nil {
+			return trace.Wrap(err, "--windows-labels")
+		}
+		allow.WindowsDesktopLabels = types.ToLabels(labels)
+	}
+	if c.windowsLoginsSet {
+		allow.WindowsDesktopLogins = utils.SplitIdentifiers(c.windowsLogins)
+	}
+
+	// GitHub
+	if c.gitHubOrgsSet {
+		orgs := utils.SplitIdentifiers(c.gitHubOrgs)
+		if len(orgs) == 0 {
+			allow.GitHubPermissions = nil
+		} else {
+			allow.GitHubPermissions = []types.GitHubPermission{{Organizations: orgs}}
+		}
+	}
+	return nil
+}
+
+// applyAWSICFlagsToRole modifies the AWS IC role's allow block.
+// Empty values clear the whole allow spec since this role is specific to awsic,
+// unset flags leave fields alone.
+func (c *Command) applyAWSICFlagsToRole(allow *types.RoleConditions) error {
+	if c.awsicAssignments == "" {
+		*allow = types.RoleConditions{}
+	} else {
+		allow.AppLabels = awsIcAppLabel
+		aa, err := buildAWSICAccountAssignments(c.awsicAssignments)
+		if err != nil {
+			return trace.Wrap(err)
+		}
+		allow.AccountAssignments = aa
+	}
+	return nil
+}
 
 func buildAWSICAccountAssignments(awsicAssignments string) ([]types.IdentityCenterAccountAssignment, error) {
 	var aa []types.IdentityCenterAccountAssignment
